Fixing the structure of the app and giving components more understandable names
