+++ src/components/BoardCard/BoardCard.styles.scss
@@ -1,4 +1,4 @@
-.board {
+.board-card {
     height: 100px;
     width: 180px;
     padding: 50px 80px;
