+++ src/components/Button/Button.styles.scss
@@ -1,0 +1,10 @@
+.button {
+    padding: 10px 25px;
+    background-color: #4169e1;
+    color: white;
+    border: 1px solid black;
+    border-radius: 5px;
+    font-size: 16px;
+    font-weight: bold;
+    cursor: pointer;
+}
