+++ src/components/SearchBar/SearchBar.component.js
