+++ src/components/SearchBar/SearchBar.styles.scss
