--- src/pages/Home.js
@@ -1,11 +1,0 @@
-import BoardContainer from "../components/BoardContainer.component";
-
-const Home = () => {
-    return (
-        <div>
-            <BoardContainer />
-        </div>
-    );
-};
-
-export default Home;
