+++ src/pages/Home.page.js
@@ -1,0 +1,11 @@
+import BoardList from "../components/BoardList/BoardList.component";
+
+const Home = () => {
+    return (
+        <div>
+            <BoardList />
+        </div>
+    );
+};
+
+export default Home;
